crop1/crop/requirements.txt
# Core packages
Flask
numpy
tensorflow-intel
Pillow
opencv-python
opencv-contrib-python
